Tests & More Tasks implemented.
We will have to discuss, whether we should continue on this path. As I see it, the introduction of Tasks seems to be redundant, and just add an extra layer of unnecessary mapping.

Would it make sense, to add the execution logic directly to the actions? (But still compile redundant actions such as mix actions into a ActionQueue etc. and not have mix as an internal action in the backend, but rather a list of moves etc.)