Fixed wrong droplet positioning in generated JSON file.
Obstacle-droplets that are mid-move are now correctly removed from available-grid (The extra electrodes taken up by these + safe area where previously not accounted for)

Droplets are now being moved by a span of electrodes that are smaller than or equal to the diameter of a droplet. This means, that a droplet that has a diameter of 25, is only being moved by a span of 1 electrode (given the electrode has a span of 20), instead of 2.

Updated DMF JSON file to remove the last remnants of the reservoirs. This includes shrinking the overall x-size + shifting all electrodes 110 to the left. (reducing positionX by 110)